mcp_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.3.27
langchain-openai>=0.3.33
python-dotenv>=1.1.1